requirements.txt
click<=8.0.4
streamlit
opencv-python
pandas
Pillow
plotly